requirements.txt
certifi==2024.7.4
beautifulsoup4==4.12.3
langchain==0.2.15
langgraph==0.2.14
langchainhub==0.1.21
langchain-community==0.2.13
langchain-openai==0.1.23
tavily-python==0.4.0
langchain-chroma==0.1.3
python-dotenv==1.0.1
pytest==8.3.2
pymongo==4.8.0
motor==3.5.1
langchain-core==0.2.35